+++ test/repository/repository_url_test.dart
@@ -83,7 +83,61 @@ void main() {
       );
     });
 
-    test('absolute reference URI', () {
+    test('absolute path', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev').resolveUrl('/y'),
+        '$prefix/dart-lang/pub-dev/blob/master/y',
+      );
+    });
+
+    test('absolute path with multiple slashes', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev').resolveUrl('/////y'),
+        '$prefix/dart-lang/pub-dev/blob/master/y',
+      );
+    });
+
+    test('relative path inside of the repository', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev/tree/master/sub/dir')
+            .resolveUrl('../y/x'),
+        '$prefix/dart-lang/pub-dev/blob/master/sub/y/x',
+      );
+    });
+
+    test('relative path outside of the repository', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev')
+            .resolveUrl('../y/../x'),
+        '$prefix/dart-lang/pub-dev/blob/master/x',
+      );
+    });
+
+    test('query parameters', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev')
+            .resolveUrl('img.png?width=800'),
+        '$prefix/dart-lang/pub-dev/blob/master/img.png?width=800',
+      );
+    });
+
+    test('relative from + relative path', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev')
+            .resolveUrl('../abcd', relativeFrom: 'example/subdir/README.md'),
+        '$prefix/dart-lang/pub-dev/blob/master/example/abcd',
+      );
+    });
+
+    test('relative from + absolute path', () {
+      expect(
+        Repository.parseUrl('$prefix/dart-lang/pub-dev')
+            .resolveUrl('/abcd', relativeFrom: 'example/subdir/README.md'),
+        '$prefix/dart-lang/pub-dev/blob/master/abcd',
+      );
+    });
+
+    test('absolute URI with scheme', () {
       expect(
         Repository.parseUrl('$prefix/dart-lang/pub-dev')
             .resolveUrl('https://example.com/y'),
@@ -98,6 +152,51 @@ void main() {
 
   group('GitLab URLs', () {
     testGitHubUrls('https://gitlab.com');
+  });
+
+  group('Unknown/homepage URLs', () {
+    final r = Repository(
+      provider: RepositoryProvider.unknown,
+      host: 'example.com',
+      repository: '/project',
+      branch: null,
+      path: null,
+    );
+
+    test('fragment link', () {
+      expect(r.resolveUrl('#x'), '#x');
+    });
+
+    test('sibling file', () {
+      expect(
+          r.resolveUrl('README.md'), 'https://example.com/project/README.md');
+    });
+
+    test('sibling file + fragment', () {
+      expect(r.resolveUrl('README.md#x'),
+          'https://example.com/project/README.md#x');
+    });
+
+    test('subdirectory', () {
+      expect(r.resolveUrl('docs'), 'https://example.com/project/docs');
+    });
+
+    test('`..` + other dir + query parameters', () {
+      expect(r.resolveUrl('../other?q=s'), 'https://example.com/other?q=s');
+    });
+
+    test('multi-level `..` + other dir', () {
+      expect(r.resolveUrl('../../../other'), 'https://example.com/other');
+    });
+
+    test('relative from + `..` + other dir', () {
+      expect(r.resolveUrl('../other', relativeFrom: 'a/b/README.md'),
+          'https://example.com/project/a/other');
+    });
+
+    test('absolute directory', () {
+      expect(r.resolveUrl('/other'), 'https://example.com/other');
+    });
   });
 
   group('URL replacements', () {